+++ src/routes/home/Home.module.css
@@ -43,7 +43,7 @@ h3 {
   position: relative;
 }
 
-.homeCategoryNav2 {
+.homeCategoryNav {
   display: flex;
   justify-content: space-between;
   list-style: none;
@@ -54,7 +54,7 @@ h3 {
   scroll-behavior: smooth;
 }
 
-.homeCategoryNav2::-webkit-scrollbar {
+.homeCategoryNav::-webkit-scrollbar {
   display: none;
 }
 
@@ -300,7 +300,7 @@ h3 {
     height: 100px;
   }
 
-  .homeCategoryNav2 {
+  .homeCategoryNav {
     grid-gap: 17px;
     margin-left: -15px;
     margin-right: -15px;
