+++ src/shared/category-nav/CategoryNav.module.css
@@ -1,0 +1,50 @@
+.container {
+  width: 100%;
+  height: 116px;
+  position: relative;
+}
+
+.navbar {
+  display: flex;
+  justify-content: space-between;
+  list-style: none;
+  overflow-y: hidden;
+  overflow-x: scroll;
+  grid-gap: 20px;
+  transition-delay: 1s;
+  scroll-behavior: smooth;
+}
+
+.navbar::-webkit-scrollbar {
+  display: none;
+}
+
+.navItem {
+  text-align: center;
+}
+
+.navItem img {
+  cursor: pointer;
+  width: 85px;
+  height: 85px;
+}
+
+
+@media screen and (max-width: 900px) {
+  .container {
+    height: 100px;
+  }
+
+  .navbar {
+    grid-gap: 17px;
+    margin-left: -15px;
+    margin-right: -15px;
+    padding: 0 15px;
+  }
+
+  .navItem img {
+    cursor: pointer;
+    width: 67px;
+    height: 67px;
+  }
+}
