--- src/shared/categoryNav/CategoryNav.module.css
@@ -1,16 +1,0 @@
-.homeCategoryContainer {
-  height: 100px;
-}
-
-.homeCategoryNav2 {
-  grid-gap: 17px;
-  margin-left: -15px;
-  margin-right: -15px;
-  padding: 0 15px;
-}
-
-.categoryCard img {
-  cursor: pointer;
-  width: 67px;
-  height: 67px;
-}
